rainy_day/requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.4
drf-yasg2==1.19.4
exceptiongroup==1.0.0rc9
factory-boy==3.2.1
Faker==15.1.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
PyPika==0.48.9
pytest==7.2.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-watch==4.2.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
watchdog==2.1.9
